requirements.txt
amqp==5.3.1
asgiref==3.9.2
billiard==4.2.2
boto3==1.40.38
botocore==1.40.38
celery==5.5.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.6
django-storages==1.14.6
djangorestframework==3.16.1
iniconfig==2.1.0
jmespath==1.0.1
kombu==5.5.4
packaging==25.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
Pygments==2.19.2
pytest==8.4.2
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==6.4.0
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
